refactor: move DockerStageRunner and CredentialsGuard into container_execution (#132)
Move DockerStageRunner, CredentialsGuard, and resume-retry logic from
run.rs into container_execution/runner.rs. The "execute one stage in
Docker" concern is now a single cohesive module. run.rs constructs
DockerStageRunner via its public constructor and passes it to
PipelineExecutor, and no longer contains any Docker execution logic.
