+++ src/container_execution/mod.rs
@@ -1,6 +1,7 @@
 mod docker_args;
 mod infra;
 mod process;
+mod runner;
 mod stream_parser;
 
 pub use docker_args::{build_docker_args, container_name_for};
@@ -8,6 +9,7 @@ pub use infra::{
     detect_compose_network, host_token_is_expired, make_mcp_config, token_remaining_minutes,
 };
 pub use process::{post_stream_error, run_container, run_iteration, StreamResult};
+pub use runner::{CredentialsGuard, DockerStageRunner};
 
 use std::path::PathBuf;
